+++ src/CMakeLists.txt
@@ -517,6 +517,7 @@ add_library(Halide ${HALIDE_LIBRARY_TYPE}
   EliminateBoolVectors.cpp
   EmulateFloat16Math.cpp
   Error.cpp
+  Expr.cpp
   FastIntegerDivide.cpp
   FindCalls.cpp
   Float16.cpp
