+++ src/IROperator.cpp
@@ -2222,14 +2222,4 @@ Expr undef(Type t) {
                                 Internal::Call::PureIntrinsic);
 }
 
-Range::Range(const Expr &min_in, const Expr &extent_in)
-    : min(lossless_cast(Int(32), min_in)), extent(lossless_cast(Int(32), extent_in)) {
-    if (min_in.defined() && !min.defined()) {
-        user_error << "Min cannot be losslessly cast to an int32: " << min_in;
-    }
-    if (extent_in.defined() && !extent.defined()) {
-        user_error << "Extent cannot be losslessly cast to an int32: " << extent_in;
-    }
-}
-
 }  // namespace Halide
